feat: add support for rendering links in ContentBody component
The `ContentBody` component now supports rendering links in the page data. This is achieved by iterating over the `page.data.links` array and rendering a `Button` component for each link. The `Button` component takes the `link` and `label` properties from the `Content.ProjectDocumentDataLinksItem` type.

This change improves the functionality of the `ContentBody` component by allowing it to display links in the rendered content.